Code improvement on master user view

+++ rental/src/rental/MasterUser.java
@@ -459,22 +459,27 @@ public class MasterUser extends javax.swing.JFrame {
      */
     private void delete(){
         try{
-            if(UserRepository.destroy(Integer.parseInt(txtId.getText()))){
-                LogRepository.create(new Log(0, authUser.getId(), "User", "DELETE"));
-                
-                JOptionPane.showMessageDialog(null, "Delete successful", "Delete", JOptionPane.INFORMATION_MESSAGE);
-                txtId.setText("");
-                txtName.setText("");
-                txtUsername.setText("");
-                txtPassword.setText("");
-                txtEmail.setText("");
-                txtPhone.setText("");
-                txtAddress.setText("");
+            int dialogButton = JOptionPane.YES_NO_OPTION;
+            int dialogResult = JOptionPane.showConfirmDialog (null, "Are you sure want to delete this record?","Delete",dialogButton);
+            
+            if(dialogResult == JOptionPane.YES_OPTION){
+                if(UserRepository.destroy(Integer.parseInt(txtId.getText()))){
+                    LogRepository.create(new Log(0, authUser.getId(), "User", "DELETE"));
 
-                getUsers();
-            }
-            else{
-                JOptionPane.showMessageDialog(null, "Oops, something's wrong", "Delete", JOptionPane.INFORMATION_MESSAGE);
+                    JOptionPane.showMessageDialog(null, "Delete successful", "Delete", JOptionPane.INFORMATION_MESSAGE);
+                    txtId.setText("");
+                    txtName.setText("");
+                    txtUsername.setText("");
+                    txtPassword.setText("");
+                    txtEmail.setText("");
+                    txtPhone.setText("");
+                    txtAddress.setText("");
+
+                    getUsers();
+                }
+                else{
+                    JOptionPane.showMessageDialog(null, "Oops, something's wrong", "Delete", JOptionPane.INFORMATION_MESSAGE);
+                }
             }
         } catch (Exception e){
             JOptionPane.showMessageDialog(null, "Please choose a record", "Delete", JOptionPane.INFORMATION_MESSAGE);
